libft files changed a bit

+++ libft/ft_split.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_split.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anovio-c <anovio-c@student.42barcel>       +#+  +:+       +#+        */
+/*   By: simarcha <simarcha@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/11 10:47:28 by anovio-c          #+#    #+#             */
-/*   Updated: 2024/02/08 13:13:55 by anovio-c         ###   ########.fr       */
+/*   Updated: 2024/06/20 17:08:33 by simarcha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ static char	**subsplit(char const *s, char c, char **arr, size_t words)
 		arr[i] = put_word(s, j, get_len_word(&s[j], c));
 		if (!arr[i])
 		{
-			while (--i >= 0)
+			while ((int)--i >= 0)
 			{
 				free(arr[i]);
 				return (NULL);
